2022-12-19 성찬민 1. 필드 추가 studyGatheredSize : 현재 스터디의 모집된 인원 2. studyJContent 자료형 변경 String -> Clob + 프론트에서 Content를 서버로 넘길 때 Clob로 변환할 것 + 데이터베이스, 백, 프론트 모두에서 Clob로 다룬다. 3. DTO BOMI 기존 DTO 로 통합 4. selectStudyByEmail의 과목정보, 과목 대분류 정보 추가 5. sselectStudy 과목 대분류 정보 추가 6.  중복 메서드 삭제 selectStudy, selectStudyByStudyId

+++ dogroupback/src/main/java/com/my/dto/StudyUserDTO.java
@@ -21,7 +21,7 @@ public class StudyUserDTO extends UserDTO {
 	public StudyUserDTO() {
 		super();
 	}
-	public StudyUserDTO(String email, String name, String password, int diligence, int userBalance, int status) {
+	public StudyUserDTO(String email, String name, String password, double diligence, int userBalance, int status) {
 		super(email, name, password, diligence, userBalance, status);
 	}
 	public int getStudyId() {
